backend/requirements.txt

fastapi
uvicorn
pydantic
python-dotenv
pillow
pytesseract
httpx
supabase
